+++ packages/bpk-component-barchart/src/BpkBarchartBar.js
@@ -32,13 +32,17 @@ const BpkBarchartBar = (props) => {
     label,
     onClick,
     outlier,
+    selected,
     ...rest
   } = props;
 
   const classNames = [getClassName('bpk-barchart__bar')];
   if (className) { classNames.push(className); }
   if (outlier) { classNames.push(getClassName('bpk-barchart__bar--outlier')); }
+  if (selected) { classNames.push(getClassName('bpk-barchart__bar--selected')); }
   if (onClick) { classNames.push(getClassName('bpk-barchart__bar--interactive')); }
+
+  const isAriaPressed = !!(onClick && selected);
 
   return (
     <g className={getClassName('bpk-barchart__bar-group')}>
@@ -55,6 +59,7 @@ const BpkBarchartBar = (props) => {
         onKeyDown={onClick ? handleKeyboardEvent(onClick) : undefined}
         tabIndex={onClick ? 0 : undefined}
         role={onClick ? 'button' : undefined}
+        aria-pressed={isAriaPressed}
         aria-label={label}
         {...rest}
       />
@@ -71,12 +76,14 @@ BpkBarchartBar.propTypes = {
   className: PropTypes.string,
   onClick: PropTypes.func,
   outlier: PropTypes.bool,
+  selected: PropTypes.bool,
 };
 
 BpkBarchartBar.defaultProps = {
   className: null,
   onClick: null,
   outlier: false,
+  selected: false,
 };
 
 export default BpkBarchartBar;